// src/rify/drag-verify/interface.ts
export interface RifyDragVerifyOptions {
  // 拼图 X 坐标
  startX: number;
  // 拼图 Y 坐标
  startY: number;
  // 拼图边长
  length: number;
  // 拼图圆形半径
  radius: number;
  // 背景图片宽度
  width: number;
  // 背景图片高度
  height: number;
  // 拼图图片地址
  src: string;
}